[JT-68] refactor: Member domain 리팩토링
