+++ build.gradle
@@ -17,7 +17,7 @@ subprojects {
     group 'shop.jtoon'
     version '0.0.1-SNAPSHOT'
 
-    apply plugin: 'java'
+    apply plugin: 'java-library'
     apply plugin: 'idea'
     apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
