+++ module-application/app-api/build.gradle
@@ -1,15 +1,23 @@
+repositories {
+    maven { url 'https://jitpack.io' }
+}
+
 dependencies {
-    // Web
+// Web
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
-    // Bean Validation
+// Bean Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
 
-    // AOP
+// JPA
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+// AOP
     implementation 'org.springframework.boot:spring-boot-starter-aop'
 
-    // Test
+// Test
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
-
+// I-AM-PORT
+    implementation 'com.github.iamport:iamport-rest-client-java:0.2.23'
 }
