--- module-application/app-api/src/main/java/shop/jtoon/member/application/T.java
@@ -1,4 +1,0 @@
-package shop.jtoon.member.application;
-
-public class T {
-}
