--- module-application/app-api/src/main/java/shop/jtoon/member/presentation/T.java
@@ -1,4 +1,0 @@
-package shop.jtoon.member.presentation;
-
-public class T {
-}
