+++ module-application/app-api/src/main/java/shop/jtoon/member/request/SignUpReq.java
@@ -6,11 +6,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import shop.jtoon.entity.Gender;
-import shop.jtoon.entity.LoginType;
-import shop.jtoon.entity.Member;
-import shop.jtoon.entity.Role;
-import shop.jtoon.request.SignUpDto;
+import shop.jtoon.dto.SignUpDto;
 
 public record SignUpReq(
 	@Pattern(regexp = EMAIL_PATTERN) String email,
@@ -21,20 +17,6 @@ public record SignUpReq(
 	@Pattern(regexp = PHONE_PATTERN) String phone,
 	@NotNull String loginType
 ) {
-
-	public Member toEntity(String encryptedPassword) {
-		return Member.builder()
-			.email(email)
-			.password(encryptedPassword)
-			.name(name)
-			.nickname(nickname)
-			.gender(Gender.from(gender))
-			.phone(phone)
-			.role(Role.USER)
-			.loginType(LoginType.from(loginType))
-			.build();
-	}
-
 	public SignUpDto toDto() {
 		return SignUpDto.builder()
 			.email(this.email)
