--- module-application/app-api/src/main/java/shop/jtoon/member/request/T.java
@@ -1,4 +1,0 @@
-package shop.jtoon.member.request;
-
-public class T {
-}
