--- module-application/app-api/src/main/java/shop/jtoon/member/response/T.java
@@ -1,4 +1,0 @@
-package shop.jtoon.member.response;
-
-public class T {
-}
