+++ module-application/build.gradle
@@ -4,7 +4,6 @@ project(':module-application:app-api') {
         implementation project(':module-domain:domain-member')
         implementation project(':module-domain:domain-payment')
         implementation project(':module-domain:domain-webtoon')
-        implementation project(':module-domain:domain-jpa')
 
         // Internal Layer
         implementation project(':module-internal:core-web')
