+++ module-domain/build.gradle
@@ -2,7 +2,7 @@ project(':module-domain:domain-member') {
     bootJar.enabled = false
 
     dependencies {
-        implementation project(':module-domain:domain-jpa')
+        api project(':module-domain:domain-jpa')
         implementation project(':module-core')
     }
 }
@@ -11,7 +11,7 @@ project(':module-domain:domain-payment') {
     bootJar.enabled = false
 
     dependencies {
-        implementation project(':module-domain:domain-jpa')
+        api project(':module-domain:domain-jpa')
         implementation project(':module-core')
     }
 }
@@ -20,7 +20,7 @@ project(':module-domain:domain-webtoon') {
     bootJar.enabled = false
 
     dependencies {
-        implementation project(':module-domain:domain-jpa')
+        api project(':module-domain:domain-jpa')
         implementation project(':module-core')
     }
 }
