+++ module-domain/domain-member/src/main/java/shop/jtoon/dto/SignUpDto.java
@@ -1,4 +1,4 @@
-package shop.jtoon.request;
+package shop.jtoon.dto;
 
 import lombok.Builder;
 import shop.jtoon.entity.Gender;
